Partially fixed completion

+++ src/readline/ft_rl_utils2.c
@@ -6,7 +6,7 @@
 /*   By: ivalimak <ivalimak@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/28 22:24:22 by ivalimak          #+#    #+#             */
-/*   Updated: 2024/04/10 13:17:50 by ivalimak         ###   ########.fr       */
+/*   Updated: 2024/04/11 11:43:52 by ivalimak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ t_rl_word	*ft_rl_strword(const char *s)
 	return (out);
 }
 
-uint8_t		ft_rl_isdir(const char *path)
+uint16_t	ft_rl_isdir(const char *path)
 {
 	struct stat	file;
